+++ lib/dom_tools.dart
@@ -10,6 +10,7 @@ export 'src/dom_tools_document.dart';
 export 'src/dom_tools_file.dart';
 export 'src/dom_tools_js.dart';
 export 'src/dom_tools_paint.dart';
+export 'src/dom_tools_scroll.dart';
 export 'src/dom_tools_touch.dart';
 export 'src/dom_tools_track.dart';
 export 'src/perspective_filter.dart';